# backend/openapi.postman.yaml
openapi: 3.0.3
info:
  title: dAstIll Backend API
  version: 0.1.0
  description: |
    Import this file into Postman as an OpenAPI definition.

    Notes:
    - `GET /api/health` is public.
    - All other `/api/*` endpoints require `x-dastill-proxy-auth`.
    - Some destructive routes additionally require `x-dastill-role: operator`.
    - SSE endpoints return `text/event-stream`.
servers:
  - url: http://localhost:3001
    description: Local backend
...
security:
  - ProxyAuth: []
paths:
...
  /api/channels/{id}/videos:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/IncludeShorts"
      - $ref: "#/components/parameters/VideoType"
      - $ref: "#/components/parameters/Acknowledged"
      - $ref: "#/components/parameters/QueueOnly"
      - $ref: "#/components/parameters/QueueTab"
    get:
      tags: [Videos]
      summary: List videos for a channel
      responses:
        "200":
          description: Videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Videos]
      summary: Get video
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        default:
          $ref: "#/components/responses/StringError"
...
components:
  securitySchemes:
    ProxyAuth:
      type: apiKey
      in: header
      name: x-dastill-proxy-auth
      description: Required for all protected endpoints.
    OperatorRole:
      type: apiKey
      in: header
      name: x-dastill-role
      description: Set to `operator` for operator-only endpoints.
    ClientIp:
      type: apiKey
      in: header
      name: x-dastill-client-ip
      description: Optional stable client identifier used for rate limiting and anonymous chat quota.
  parameters:
    ChannelId:
      name: id
      in: path
      required: true
      schema:
        type: string
    VideoId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    IncludeShorts:
      name: include_shorts
      in: query
      schema:
        type: boolean
        default: true
    VideoType:
      name: video_type
      in: query
      schema:
        $ref: "#/components/schemas/VideoTypeFilter"
    Acknowledged:
      name: acknowledged
      in: query
      schema:
        type: boolean
    QueueOnly:
      name: queue_only
      in: query
      schema:
        type: boolean
    QueueTab:
      name: queue_tab
      in: query
      schema:
        $ref: "#/components/schemas/QueueTab"
  responses:
    StringError:
      description: Error message returned as plain text
      content:
        text/plain:
          schema:
            type: string
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: string
      additionalProperties: true
      example:
        error: Something went wrong
    ContentStatus:
      type: string
      enum: [pending, loading, ready, failed]
...
    VideoTypeFilter:
      type: string
      enum: [all, long, short]
    QueueTab:
      type: string
      enum: [transcripts, summaries, evaluations]
...
    Video:
      type: object
      required:
        - id
        - channel_id
        - title
        - published_at
        - is_short
        - transcript_status
        - summary_status
        - acknowledged
        - retry_count
      properties:
        id:
          type: string
        channel_id:
          type: string
        title:
          type: string
        thumbnail_url:
          type: string
          nullable: true
        published_at:
          type: string
          format: date-time
        is_short:
          type: boolean
        transcript_status:
          $ref: "#/components/schemas/ContentStatus"
        summary_status:
          $ref: "#/components/schemas/ContentStatus"
        acknowledged:
          type: boolean
        retry_count:
          type: integer
          minimum: 0
          maximum: 255
        quality_score:
          type: integer
          minimum: 0
          maximum: 255
          nullable: true
...